Fix nav/arith combination in loops
If a loop was the final instruction in some sequence, then it was not
entered into for combination optimisations. This is because the
combination would only be attempted if the current instruction had a
non-NULL .next element. This makes sense for nav/arith, since there's
never anything to merge if it's the last instruction in a sequence, but
not for loops as we should always enter into them.